Move encoder require up the stack
This way we require msgpack when calling
`ActiveSupport::EventReporter.encoder(:msgpack)` which can be setup and
stored during boot, rather than when calling `Encoder#encoder` event.

In this way it is slightly earlier to detect, for example, missing gem
dependency without the need for a config and railtie initializer.

We could still go that route though, just fleshing out this idea first.
